Added character achievements
Added support for retrieving character achievements.  Added a flags enum
and helper class to construct the querystring parameter to specify which
optional character fields to return.  Added a custom Json.NET contract
resolver to handle the conversion of Unix epoch numbers to DateTime
without requiring an attribute on each property.