Added integration tests
Added a unit test project that really contains xUnit-based integration
tests that call the Blizzard Community API.  And do very little
verification of the returned results.  To run the tests, you must
acquire your own API key and store it in an appSettings.json file.  An
appSettingsTemplate.json file is included as a template.  Updated the
.gitignore to ignore the appSettings.json file with the API key.