Use a single instance of HttpClient
HttpClient is intended to be instantiated once and re-used throughout the life of an application.
Especially in server applications, creating a new HttpClient instance for every request will exhaust the
number of sockets available under heavy loads. This will result in SocketException errors.

To reslove this issue I created a singleton class and everytime you create a new WarcraftClient it will grab the current HttpClient instasnce.