Updated WarcraftClient to capture error response
Updated WarcraftClient to capture error response content when a
non-success status code is returned from the World of Warcraft Community
APIs.  The details of the request and response are included in the
message of a WarcraftClientException that is thrown.  API keys in the
request URI are masked for security purposes.  An example error message
looks like the following:

Response code 404 (Not Found) does not indicate success. Request:
https://us.api.battle.net/wow/spell/8056?locale=en_US&apikey={apiKey}
Response: {"status":"nok", "reason": "unable to get spell information."}