Start region enum members at 1
By default enum members will start at 0. If you were to parse `Region` with an invaild name/value it will return `China`. This is because the default value is the enum member 0.

However if the `Region` enum members started at 1 and you parsed with an invaild name/value it will literally return 0. Although its type will remain `Region` (enum parsing is werid in C# go figure).

So by starting enum member values as 1 we can determine after parsing if the enum is vaild or not. This is very useful if you're taking a region input from an end-user such as a website.