Add better error handling
The previous method of error handling threw an exception if the http request was not successful. This solution worked but meant if you wanted to capture the error you’d have to wrap the call in a try catch statement. Furthermore, it’s better practice to eliminate the need for throwing exceptions and that’s what I’ve done. The problem posed was that there were two potential return types, the error or the object e.g. a character. Then I stumbled upon a new upcoming feature of ASP.NET Core 2.1. A new action result called ‘ActionResult<T>’. With this new type you can do the following…

    public ActionResult<Product> GetProduct(int? productId)
    {
        if (productId == null)
        {
            return BadRequest();
        }

        return new Product { SampleInt = productId.Value, };
    }

As you can see the method returns two types, a status code or type ‘T’. So I more or less copied the code that Microsoft had written for this found here: https://github.com/aspnet/Mvc/blob/dev/src/Microsoft.AspNetCore.Mvc.Core/ActionResultOfT.cs
When making a call with the API you can now do it like this…

public IActionResult Index()
{
    RequestResult<Zone> result = await _warcraftClient.GetZoneAsync(34);
    if (!result.Success)
    {
        ModelState.AddModelError(result.Error.Detail);

        return View();
    }

    return View(result.Value);
}