Moved all existing models into CommunityApi and Enumerations subfolders
Moved all existing classes in the Model folder into a CommunityApi subfolder.  Moved enumerations in the Models folder to an Enumerations subfolder.  This will lay the groundwork for dealing distinguishing between models like Achievement that will differ in structure between the Community API and the Game Data API or Profile API.