Added Character Profile Summary for Character Profile API
Added support for the Profile API with Character Profile Summary and Character Profile Status.  Added an IWarcraftClientCharacterProfileApi interface to describe the capabilities of the API.  The IWarcraftClient interface implements this interface via the IWarcraftClientProfileApi interface.  Also updated the sample code in README.md to use methods for the new Profile API.