Added README.md files for NuGet packages
Added README.md files for the ArgentPonyWarcraftClient and ArgentPonyWarcraftClient.Extensions.DependencyInjection NuGet packages.  These are derived from the primary README.md for the repository and the Docusaurus documentation for dependency injection.  Moved the ArgentPony.png graphic into doc subfolders of each project to make it easier to someday use a different graphic for each package if desired, and to enable simple relative paths.  Moved the graphic reference out of the shared Directory.Build.props file.