Added AssetWithoutFileDataId for CreatureDisplayMedia
Added a new AssetWithoutFileDataId class and updated the Assets property of CreatureDisplayMedia to use it since the non-nullable FileDataId does not appear to be returned with for this type.  This is a small change, but a breaking change, so I'm also updating the version.json to bump the major version.